Fix Bug and move get AQ
